Make StringMapEntryChange into sealed class so that serializer can serialize it without extra information
