+++ app/src/main/java/de/westnordost/streetcomplete/data/osm/edits/update_tags/StringMapEntryChange.kt
@@ -1,10 +1,41 @@
 package de.westnordost.streetcomplete.data.osm.edits.update_tags
 
-// TODO make sealed as soon as Kotlin supports that (1.5)
-interface StringMapEntryChange {
-    override fun toString(): String
-    override fun equals(other: Any?): Boolean
-    fun conflictsWith(map: Map<String, String>): Boolean
-    fun applyTo(map: MutableMap<String, String>)
-    fun reversed(): StringMapEntryChange
+import kotlinx.serialization.Serializable
+
+// TODO make into sealed INTERFACE as soon as Kotlin supports that (1.5)
+@Serializable
+sealed class StringMapEntryChange {
+    abstract override fun toString(): String
+    abstract override fun equals(other: Any?): Boolean
+    abstract override fun hashCode(): Int
+    abstract fun conflictsWith(map: Map<String, String>): Boolean
+    abstract fun applyTo(map: MutableMap<String, String>)
+    abstract fun reversed(): StringMapEntryChange
+}
+
+@Serializable
+data class StringMapEntryAdd(val key: String, val value: String) : StringMapEntryChange() {
+
+    override fun toString() = "ADD \"$key\"=\"$value\""
+    override fun conflictsWith(map: Map<String, String>) = map.containsKey(key) && map[key] != value
+    override fun applyTo(map: MutableMap<String, String>) { map[key] = value }
+    override fun reversed() = StringMapEntryDelete(key, value)
+}
+
+@Serializable
+data class StringMapEntryModify(val key: String, val valueBefore: String, val value: String) : StringMapEntryChange() {
+
+    override fun toString() = "MODIFY \"$key\"=\"$valueBefore\" -> \"$key\"=\"$value\""
+    override fun conflictsWith(map: Map<String, String>) = map[key] != valueBefore && map[key] != value
+    override fun applyTo(map: MutableMap<String, String>) { map[key] = value }
+    override fun reversed() = StringMapEntryModify(key, value, valueBefore)
+}
+
+@Serializable
+data class StringMapEntryDelete(val key: String, val valueBefore: String) : StringMapEntryChange() {
+
+    override fun toString() = "DELETE \"$key\"=\"$valueBefore\""
+    override fun conflictsWith(map: Map<String, String>) = map.containsKey(key) && map[key] != valueBefore
+    override fun applyTo(map: MutableMap<String, String>) { map.remove(key) }
+    override fun reversed() = StringMapEntryAdd(key, valueBefore)
 }
