--- app/src/main/java/de/westnordost/streetcomplete/data/osm/edits/update_tags/StringMapEntryDelete.kt
@@ -1,12 +1,0 @@
-package de.westnordost.streetcomplete.data.osm.edits.update_tags
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class StringMapEntryDelete(val key: String, val valueBefore: String) : StringMapEntryChange {
-
-    override fun toString() = "DELETE \"$key\"=\"$valueBefore\""
-    override fun conflictsWith(map: Map<String, String>) = map.containsKey(key) && map[key] != valueBefore
-    override fun applyTo(map: MutableMap<String, String>) { map.remove(key) }
-    override fun reversed() = StringMapEntryAdd(key, valueBefore)
-}
