--- app/src/main/java/de/westnordost/streetcomplete/data/osm/edits/update_tags/StringMapEntryModify.kt
@@ -1,12 +1,0 @@
-package de.westnordost.streetcomplete.data.osm.edits.update_tags
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class StringMapEntryModify(val key: String, val valueBefore: String, val value: String) : StringMapEntryChange {
-
-    override fun toString() = "MODIFY \"$key\"=\"$valueBefore\" -> \"$key\"=\"$value\""
-    override fun conflictsWith(map: Map<String, String>) = map[key] != valueBefore && map[key] != value
-    override fun applyTo(map: MutableMap<String, String>) { map[key] = value }
-    override fun reversed() = StringMapEntryModify(key, value, valueBefore)
-}
